Moved dependencies to actual package, made image_describer an installable package, minor changes in docs.
