+++ pyproject.toml
@@ -10,15 +10,11 @@ dependencies = [
     "ipdb>=0.13.13",
     "linkify-it-py>=2.0.3",
     "myst-parser>=4.0.1",
-    "pillow>=11.1.0",
     "pre-commit>=4.1.0",
     "pytest>=8.3.4",
     "ruff>=0.9.3",
     "sphinx>=8.1.3",
     "sphinx-autodoc-typehints>=3.0.1",
-    "torch>=2.6.0",
-    "torchvision>=0.21.0",
-    "transformers>=4.49.0",
 ]
 
 # Tools config
@@ -32,4 +28,7 @@ skip-magic-trailing-comma = false
 line-ending = "auto"
 
 [tool.uv.workspace]
-members = ["training/python/packaging_example"]
+members = [
+    "src/python/packaging_example",
+    "src/python/image_describer"
+]
