+++ src/docs/source/generate_index.py
@@ -1,14 +1,9 @@
 from pathlib import Path
 import os
 
-REPO_ROOT = Path(__file__).resolve().parents[2]
-DOCS_INDEX = REPO_ROOT / "docs/source/index.md"
+REPO_ROOT = Path(__file__).resolve().parents[3]
+DOCS_INDEX = REPO_ROOT / "src/docs/source/index.md"
 DOCS_SOURCE = DOCS_INDEX.parent
-
-
-def find_level(readme: Path) -> int:
-    """Find the level of the README file based on its directory depth."""
-    return len(readme.relative_to(REPO_ROOT).parts) - 1
 
 
 def generate_index():
@@ -17,8 +12,6 @@ def generate_index():
         if any(p.startswith(".") for p in readme.resolve().parts):
             continue
         rel_path = os.path.relpath(readme, start=DOCS_SOURCE)
-        level = find_level(readme)
-        lines.append(f"#{'#' * level} {readme.parent.name}")
         lines.append(f"```{{include}} {rel_path}")
         lines.append("```")
         lines.append("")
