+++ src/docs/source/index.md
@@ -1,37 +1,28 @@
 # Project Docs
 
-# dev-playground
-```{include} ../../README.md
+```{include} ../../../README.md
 ```
 
-## bazel
 ```{include} ../../bazel/README.md
 ```
 
-## cpp
 ```{include} ../../cpp/README.md
 ```
 
-## github
 ```{include} ../../github/README.md
 ```
 
-## python
 ```{include} ../../python/README.md
 ```
 
-### image_describer
 ```{include} ../../python/image_describer/README.md
 ```
 
-### packaging_example
 ```{include} ../../python/packaging_example/README.md
 ```
 
-### standalone_scripts
 ```{include} ../../python/standalone_scripts/README.md
 ```
 
-## uv
 ```{include} ../../uv/README.md
 ```
