+++ src/python/image_describer/src/image_describer/describer.py
